// wishinsight-be-master/enums/enums.ts
export enum Collection {
  Users = "Users",
  JobListings = "JobListings",
  Companies = "Companies",
  ApprovedProjects = "ApprovedProjects",
  AdminTokens = "AdminTokens",
  Metadata = "Metadata",
  RejectedProjects = "RejectedProjects",
  Newsletter_Signups = "Newsletter_Signups",
  UserTokens = "UserTokens",
  ChatRooms = "ChatRooms",
  ChatMessages = "ChatMessages",
}

export enum ItemStatuses {
  Approved = "approved",
  Pending = "pending",
  Rejected = "rejected",
  Filled = "filled",
}

export enum UserCollections {
  Items = "Items",
  SellingItems = "SellingItems",
}

export enum ItemTypes {
  Niche = "niche",
  Bargain = "bargain",
}